requirements.txt
altgraph==0.17.4
archspec @ file:///home/conda/feedstock_root/build_artifacts/archspec_1708969572489/work
bcrypt==4.2.1
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1733827268945/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1725267521616/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560565029/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1733218092148/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
conda @ file:///Users/runner/miniforge3/conda-bld/conda_1732643645977/work
conda-libmamba-solver @ file:///home/conda/feedstock_root/build_artifacts/conda-libmamba-solver_1727359833193/work/src
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1733855758490/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1729004031731/work
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1704321475663/work
docutils==0.21.2
fonttools==4.55.3
frozendict @ file:///Users/runner/miniforge3/conda-bld/frozendict_1728841357330/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1733298745555/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1733299205993/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1733298771451/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1733814567314/work
jsonpointer @ file:///Users/runner/miniforge3/conda-bld/jsonpointer_1725302889955/work
Kivy==2.3.0
Kivy-Garden==0.1.5
kiwisolver==1.4.7
libmambapy @ file:///Users/runner/miniforge3/conda-bld/mamba-split_1732267694058/work/libmambapy
macholib==1.16.3
mamba @ file:///Users/runner/miniforge3/conda-bld/mamba-split_1732267694058/work/mamba
matplotlib==3.10.0
menuinst @ file:///Users/runner/miniforge3/conda-bld/menuinst_1731147002256/work
modulegraph==0.19.6
numpy==2.2.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas==2.2.3
paramiko==3.5.0
pdf2image==1.17.0
pillow==11.0.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1733232627818/work
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1733222765875/work
psutil==6.1.1
py2app==0.28.8
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1732588447334/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PyNaCl==1.5.0
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
pytz==2024.2
qrcode==8.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
ruamel.yaml @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml_1728764995169/work
ruamel.yaml.clib @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml.clib_1728724470051/work
seaborn==0.13.2
setuptools==75.6.0
six==1.17.0
tk==0.1.0
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1732497199771/work
truststore @ file:///home/conda/feedstock_root/build_artifacts/truststore_1729762363021/work
tzdata==2024.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1733206856906/work
wheel==0.45.1
yt-dlp==2024.12.23
zstandard==0.23.0
